+++ src/reader/async_tokio.rs
@@ -529,11 +529,11 @@ impl<R: AsyncBufRead + Unpin> NsReader<R> {
     ///     match reader.read_resolved_event_into_async(&mut buf).await.unwrap() {
     ///         (Bound(Namespace("www.xxxx")), Event::Start(e)) => {
     ///             count += 1;
-    ///             assert_eq!(e.local_name(), QName("tag1").into());
+    ///             assert_eq!(e.local_name(), "tag1");
     ///         }
     ///         (Bound(Namespace("www.yyyy")), Event::Start(e)) => {
     ///             count += 1;
-    ///             assert_eq!(e.local_name(), QName("tag2").into());
+    ///             assert_eq!(e.local_name(), "tag2");
     ///         }
     ///         (_, Event::Start(_)) => unreachable!(),
     ///